--- packages/york-analytics/src/components/AnalyticsContext/README.MD
@@ -1,3 +1,0 @@
-```js static
-const { Provider, Consumer } = AnalyticsContext
-```
