+++ packages/york-analytics/src/index.js
@@ -1,3 +1,3 @@
+export { default as AnalyticsContext } from './context'
 export { default as AnalyticsProvider } from './components/AnalyticsProvider'
-export { default as AnalyticsContext } from './components/AnalyticsContext'
 export { default as useAnalytics } from './hooks/useAnalytics'
